clang warning fixes
Fixes clang's sign conversion warnings.

+++ tests/test_python_types.cpp
@@ -461,8 +461,8 @@ test_initializer python_types([](py::module &m) {
 
     m.attr("wchar_size") = py::cast(sizeof(wchar_t));
     m.def("ord_char", [](char c) -> int { return static_cast<unsigned char>(c); });
-    m.def("ord_char16", [](char16_t c) -> int { return c; });
-    m.def("ord_char32", [](char32_t c) -> int { return c; });
+    m.def("ord_char16", [](char16_t c) -> uint16_t { return c; });
+    m.def("ord_char32", [](char32_t c) -> uint32_t { return c; });
     m.def("ord_wchar", [](wchar_t c) -> int { return c; });
 });
